Begin moving to typescript
At this point, the only typescript module we have is `$studio/index.tsx`. As time goes on,
we'll convert more modules to typescript.

Thankfully, tons of TypeScript's benefits are already usable at this stage. Autocompletion
works well, so do refactorings, autofixes, auto import, and goto definition. So let's not
fret about converting all of our flow code to typescript asap to get full type coverage.

Also, while it was a goal with Flow to get as much type coverage as possible (it was supposed
let us write all of that DataVerse atoms while still giving us the inellisense of regular JS
values – it didn't), we know for a fact that neither flow nor typescript are
powerful enough at this stage to do the complex typelevel programming required for that kind
of type coverage. So, now the recommended guideline is more in the lines of let's get 80% of
the benefits of the typesystem by putting in 20% of the effort, and let's just forget about
the extra 20. (You can choose not to follow this recommendation ofc)