Make default and static overrides distinguishable
Right now it's easy for the user to distinguish between a sequenced prop and non-sequenced props.

However, among the non-sequenced ones, it's not possible to distinguish between those that have a static override, and those that don't. The goal of this commit is to make it easy to distinguish between the two.